+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ## 1.5 branch
 ### 1.5.10
-* `Exceptionist::__callStatic()` now handles calls starting with "isNot";
+* `Exceptionist::__callStatic()` now handles calls containing with the "Not"
+    word (e.g. `isNotArray()` or `fileNotExists()`);
 * added `Exceptionist::isFalse()` method;
 * many methods already supported by `Exceptionist` have been documented.